+++ prowler/providers/gcp/gcp_provider.py
@@ -86,6 +86,7 @@ class GcpProvider(Provider):
         client_secret: str = None,
         refresh_token: str = None,
         service_account_key: dict = None,
+        skip_api_check: bool = False,
     ):
         """
         GCP Provider constructor
@@ -107,6 +108,7 @@ class GcpProvider(Provider):
             client_secret: str
             refresh_token: str
             service_account_key: dict
+            skip_api_check: bool
 
         Raises:
             GCPNoAccesibleProjectsError if no project IDs can be accessed via Google Credentials
@@ -173,6 +175,10 @@ class GcpProvider(Provider):
 
             gcp_config.DEFAULT_RETRY_ATTEMPTS = retries_max_attempts
             logger.info(f"GCP retry attempts set to {retries_max_attempts}")
+
+        if skip_api_check:
+            logger.info("Skipping API active check for each service")
+        self.skip_api_check = skip_api_check
 
         self._impersonated_service_account = impersonate_service_account
         # Set the GCP credentials using the provided client_id, client_secret and refresh_token